3DFACE: corrections in member types of dxf_3dface_read.

+++ src/3dface.h
@@ -1,8 +1,8 @@
 /*!
  * \file 3dface.h
  *
- * \author Copyright (C) 2010, 2012, 2013, 2014, 2015 2016, 2017, 2018
- * by Bert Timmerman <bert.timmerman@xs4all.nl>.
+ * \author Copyright (C) 2010, 2012, 2013, 2014, 2015 2016, 2017, 2018,
+ * 2019 by Bert Timmerman <bert.timmerman@xs4all.nl>.
  *
  * \brief Header file for a DXF 3D face entity (\c 3DFACE).
  *
@@ -95,20 +95,20 @@ dxf_3dface_struct
                  * </ol>
                  * Group code = 60.\n
                  * \since Introduced in version R13. */
-        int color;
+        int16_t color;
                 /*!< Color of the entity.\n
                  * Defaults to \c BYLAYER if ommitted in the DXF file.\n
                  * Note that entities encapsulated in a block with the
                  * color \c BYBLOCK are represented in the "native" color of
                  * the \c BLOCK entity.\n
                  * Group code = 62. */
-        int paperspace;
+        int16_t paperspace;
                 /*!< Entities are to be drawn on either \c PAPERSPACE or
                  * \c MODELSPACE.\n
                  * Optional, defaults to \c DXF_MODELSPACE (0).\n
                  * Group code = 67.\n
                  * \since Introduced in version R13. */
-        int graphics_data_size;
+        int32_t graphics_data_size;
                 /*!< Number of bytes in the proxy entity graphics
                  * represented in the sub-sequent 310 groups, which are
                  * binary chunk records (optional).\n
@@ -236,12 +236,12 @@ double dxf_3dface_get_linetype_scale (Dxf3dface *face);
 Dxf3dface *dxf_3dface_set_linetype_scale (Dxf3dface *face, double linetype_scale);
 int16_t dxf_3dface_get_visibility (Dxf3dface *face);
 Dxf3dface *dxf_3dface_set_visibility (Dxf3dface *face, int16_t visibility);
-int dxf_3dface_get_color (Dxf3dface *face);
-Dxf3dface *dxf_3dface_set_color (Dxf3dface *face, int color);
-int dxf_3dface_get_paperspace (Dxf3dface *face);
-Dxf3dface *dxf_3dface_set_paperspace (Dxf3dface *face, int paperspace);
-int dxf_3dface_get_graphics_data_size (Dxf3dface *face);
-Dxf3dface *dxf_3dface_set_graphics_data_size (Dxf3dface *face, int graphics_data_size);
+int16_t dxf_3dface_get_color (Dxf3dface *face);
+Dxf3dface *dxf_3dface_set_color (Dxf3dface *face, int16_t color);
+int16_t dxf_3dface_get_paperspace (Dxf3dface *face);
+Dxf3dface *dxf_3dface_set_paperspace (Dxf3dface *face, int16_t paperspace);
+int32_t dxf_3dface_get_graphics_data_size (Dxf3dface *face);
+Dxf3dface *dxf_3dface_set_graphics_data_size (Dxf3dface *face, int32_t graphics_data_size);
 int16_t dxf_3dface_get_shadow_mode (Dxf3dface *face);
 Dxf3dface *dxf_3dface_set_shadow_mode (Dxf3dface *face, int16_t shadow_mode);
 DxfBinaryGraphicsData *dxf_3dface_get_binary_graphics_data (Dxf3dface *face);
